requirements.txt
amqp==5.0.6
asgiref==3.3.4
backcall==0.2.0
billiard==3.6.4.0
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
cssselect2==0.4.1
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.4
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-simple-history==3.0.0
django-weasyprint==1.1.0
django-widget-tweaks==1.4.8
Faker==8.11.0
flake8==3.9.2
flake8-builtins==1.5.3
flake8-django==1.1.2
flake8-import-order==0.18.1
flake8-print==4.0.0
html5lib==1.1
idna==2.10
ipdb==0.13.9
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
kombu==5.1.0
Markdown==3.3.4
matplotlib-inline==0.1.2
mccabe==0.6.1
oauthlib==3.1.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.1.0
pyphen==0.11.0
python-dateutil==2.8.2
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
tinycss2==1.1.0
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==53.0
webencodings==0.5.1
